Assets/omnidata-main/omnidata_tools/torch/requirements.txt
numpy
pillow
pyyaml
torch
torchvision
pytorch_lightning
einops
matplotlib
joblib
pandas
h5py
scipy
seaborn
kornia
timm